+++ runnable/generate.js
@@ -6,18 +6,25 @@ import fs   from 'fs'
 
 const input = fs.readFileSync(path.join(__dirname, process.argv[2]), 'utf8')
 
+// Included countries
 let included_countries
-
 if (process.argv[3])
 {
 	included_countries = process.argv[3].split(',')
-	console.log('Included countries:')
-	console.log(included_countries)
+	console.log('Included countries:', included_countries)
 	included_countries = new Set(included_countries)
 }
 
+// Include all regular expressions
+let extended = false
+if (process.argv[4] === 'extended')
+{
+	console.log('Include extra validation regular expressions')
+	extended = true
+}
+
 // Generate and compress metadata
-generate(input, included_countries).then((output) =>
+generate(input, included_countries, extended).then((output) =>
 {
 	// Write uncompressed metadata into a file for easier debugging
 	fs.writeFileSync(path.join(__dirname, '../metadata.json'), JSON.stringify(output, undefined, 3))