+++ source/validate.js
@@ -1,6 +1,29 @@
-import parse from './parse'
+import parse, { get_number_type } from './parse'
+
+import
+{
+	get_types
+}
+from './metadata'
 
 export default function is_valid(number, country_code)
 {
-	return Object.keys(parse.call(this, number, country_code)).length > 0
+	const parsed = parse.call(this, number, country_code)
+
+	if (!parsed.country)
+	{
+		return false
+	}
+
+	const country_metadata = this.metadata.countries[parsed.country]
+
+	if (get_types(country_metadata))
+	{
+		if (!get_number_type(parsed.phone, country_metadata))
+		{
+			return false
+		}
+	}
+
+	return true
 }